size improvements

+++ runFromWeb/Shortcut.ps1
@@ -1,6 +1,14 @@
 $fileLocation = "$env:TEMP\CCStopperDownloader.ps1"
 
-Invoke-RestMethod "https://raw.githubusercontent.com/eaaasun/CCStopper/dev/runFromWeb/Downloader.ps1" -UseBasicParsing | Out-File $fileLocation -Force
+if (Test-Path -Path "$env:USERPROFILE\Desktop\CCStopper (Online).lnk") {
+	Write-Host "Shortcut already exists!"
+	Pause
+	exit
+}
+else {
+	Invoke-RestMethod "https://raw.githubusercontent.com/eaaasun/CCStopper/dev/runFromWeb/Downloader.ps1" -UseBasicParsing | Out-File $fileLocation -Force
 
-Start-Process -FilePath "powershell.exe" -ArgumentList "-NoProfile -ExecutionPolicy Bypass -File `"$fileLocation`" -shortcut" -Verb runas -Wait
-Remove-Item -Path $fileLocation -Force
+	Start-Process -FilePath "powershell.exe" -ArgumentList "-NoProfile -ExecutionPolicy Bypass -File `"$fileLocation`" -shortcut" -Verb runas -Wait
+	Remove-Item -Path $fileLocation -Force
+	
+}
